+++ app/controllers/posts_controller.rb
@@ -13,7 +13,7 @@ class PostsController < ApplicationController
 
   def create
     @topic = Topic.find_by(id: params[:topic_id])
-    @post = Post.new(post_params.merge(topic_id: params[:topic_id]))
+    @post = current_user.posts.build(post_params.merge(topic_id: params[:topic_id]))
 
     if @post.save
       flash[:success] = "You've created a new post."
@@ -27,6 +27,7 @@ class PostsController < ApplicationController
   def edit
     @post = Post.find_by(id: params[:id])
     @topic = @post.topic
+    authorize @post
   end
 
   def update
@@ -34,8 +35,10 @@ class PostsController < ApplicationController
     @post = Post.find_by(id: params[:id])
 
     if @post.update(post_params)
+      flash[:success] = "You've edited the post."
       redirect_to topic_posts_path(@topic)
     else
+      flash[:danger] = @post.errors.full_messages
       redirect_to edit_topic_post_path(@topic, @post)
     end
   end
@@ -43,8 +46,10 @@ class PostsController < ApplicationController
   def destroy
     @post = Post.find_by(id: params[:id])
     @topic = @post.topic
+    authorize @post
 
     if @post.destroy
+      flash[:success] = "You've deleted the post."
       redirect_to topic_posts_path(@topic)
     # else
     #   redirect_to post_path(@post)
